Rename: style -> comment.style.js
